Move detailing funcs to separate file
